api/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
mysql-connector-python==9.2.0
pydantic==2.10.6
